Added some functionality to Store.java for handling purchase requests from Store.jsp. Created StoreHelper.java which is a helper bean for handling purchase requests.
Added a test login to PersistenceManager.java to aid in testing the application. If you log in with a user name of test and a password of test, PersistenceManager will create a user obect without going to the database. This will allow testing without having to set up a db.

Added some more potentially useful constants to Constants.java