Added NavPageHeader to Market page.
Started adding timer to Fruits.java. Each time a Fruits object is created (user plants a fruit in the game) a thread will start with  a timer. When the timer expires the fruit is ready.

Created FruitsEnum. This defines fruit names, base costs, and the time required for the fruit to grow.

Created QualitiesEnum. This defines quality types and a cost coefficient for each quality type.