Added Utils.java. Right now this just has one method which will get a User object from a session.
Garden.java: added arraylist of fruits. Each User object will have a Garden object associated with it. The Garden object will contain all of the fruits that the user has.

Added timeToRot to FruitsEnum. This field specifies how long each fruit takes to rot after it grows.

Added getQuality method to QualitiesEnum. This method can be used to determine the quality of a fruit based on number of waterings and fertilizers.

In Fruits.java changed type to FruitsEnum instead of string. Added quality field which is a QualitiesEnum. Added isGrown boolean. Added startGrowing method which will begin a timer. When the timer expires the run method will be called.
In the run method another timer is set for the fruit rotting.